# More efficient parsing of floats and doubles