Tested with Scala 3 and JDK-21 on Intel® Core™ i7-11800H:

#### Before
```txt
Benchmark                          (size)   Mode  Cnt        Score       Error  Units
ArrayOfDoublesReading.zioJson         128  thrpt    5   135920.917 ±  2255.318  ops/s
ArrayOfFloatsReading.zioJson          128  thrpt    5   213677.601 ±  5056.059  ops/s
GeoJSONReading.zioJson                N/A  thrpt    5     9315.513 ±   271.037  ops/s
PrimitivesReading.zioJson             N/A  thrpt    5  3515039.119 ± 21043.374  ops/s
```

#### After
```txt
Benchmark                          (size)   Mode  Cnt        Score       Error  Units
ArrayOfDoublesReading.zioJson         128  thrpt    5   149040.245 ±  1606.733  ops/s
ArrayOfFloatsReading.zioJson          128  thrpt    5   233975.363 ± 28846.808  ops/s
GeoJSONReading.zioJson                N/A  thrpt    5    10582.132 ±   139.006  ops/s
PrimitivesReading.zioJson             N/A  thrpt    5  3618949.962 ± 74998.630  ops/s
```